pyproject.toml
[tool.poetry]
name = "graphql-relay"
version = "3.0.0a2"
description = """
Relay library for graphql-core-next"""
license="MIT"
authors = [
    "Syrus Akbary <me@syrusakbary.com>",
    "Christoph Zwerschke <cito@online.de>"
]
readme = "README.md"
homepage = "https://github.com/graphql-python/graphql-relay-py"
repository = "https://github.com/graphql-python/graphql-relay-py"
documentation = "https://github.com/graphql-python/graphql-relay-py"
keywords = ["graphql", "relay", "api"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8"
]
packages = [
    { include = "graphql_relay", from = "src" },
    { include = "tests", format = "sdist" },
]

[tool.poetry.dependencies]
python = "^3.6"
graphql-core = "^3.0.0a0"

[tool.poetry.dev-dependencies]
pytest = "^5.3"
pytest-asyncio = ">=0.10,<1"
pytest-cov = "^2.8,<3"
pytest-describe = ">=0.12,<1"
pyyaml = "^5.2"
black = "19.10b0"
flake8 = "^3.7"
mypy = ">=0.761,<0.770"
codecov = "^2"
check-manifest = ">=0.40,<1"
bump2version = ">=0.5,<1"
tox = "^3.14"

[tool.black]
target-version = ['py36', 'py37', 'py38']

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"
